Break out middleware to individual files
