+++ src/middleware/index.ts
@@ -1,12 +1,3 @@
-export {
-	txErrorMiddleware as txError,
-	httpErrorMiddleware as httpError,
-	errorMiddleware as error,
-	legacyErrorMiddleware as legacyError,
-	internalErrorMiddleware as internalError,
-} from './error_middleware';
-export {
-	productionLoggerMiddleware as productionLogger,
-	developmentLoggerMiddleware as developmentLogger,
-} from './logger_middleware';
-export { validateAddressMiddleware as validateAddress } from './validations_middleware';
+export * from './error';
+export * from './logger';
+export * from './validate';
