+++ src/middleware/logger/index.ts
@@ -1,0 +1,2 @@
+export { errorLoggerMiddleware as errorLogger } from './errorLoggerMiddleware';
+export { allLoggerMiddleware as allLogger } from './allLoggerMiddleware';
