+++ src/middleware/validate/index.ts
@@ -1,0 +1,1 @@
+export { validateAddressMiddleware as validateAddress } from './validateAddressMiddleware';
